requirements.txt
yt-dlp
openai-whisper
googletrans==4.0.0-rc1
torch
google-generativeai
python-dotenv
